requirements.txt
langchain-groq
langchain-ollama
langchain-core
streamlit
python-dotenv
langchain
langchain-community
pdfplumber